# doc/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusHub V1 API",
    "version": "1.0.0-design",
    "description": "开发前契约，尚无运行服务。全部业务POST，只有CSRF引导为GET。跨字段领域规则见模块文档。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
...
  "paths": {
...
    "/api/analysis/tasks/create": {
      "post": {
        "operationId": "createSolveTask",
        "tags": [
          "analysis"
        ],
        "summary": "创建分析",
        "description": "权限：analysis:run + 来源可读。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveTaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolveCreateRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/analysis/tasks/search": {
      "post": {
        "operationId": "searchSolveTasks",
        "tags": [
          "analysis"
        ],
        "summary": "我的分析记录",
        "description": "权限：analysis:run + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSearchRequest"
              }
            }
          }
        }
      }
    },
    "/api/analysis/tasks/detail": {
      "post": {
        "operationId": "solveTaskDetail",
        "tags": [
          "analysis"
        ],
        "summary": "分析详情",
        "description": "权限：analysis:run + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveTaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdRequest"
              }
            }
          }
        }
      }
    },
    "/api/analysis/tasks/cancel": {
      "post": {
        "operationId": "cancelSolveTask",
        "tags": [
          "analysis"
        ],
        "summary": "取消任务",
        "description": "权限：analysis:run + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveTaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "NEXUSHUB_SESSION"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
...
      "GameState": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schemaVersion": {
            "const": 1,
            "type": "integer"
          },
          "rulesetVersion": {
            "type": "string",
            "enum": [
              "CLASSIC_V1"
            ]
          },
          "hands": {
            "type": "array",
            "items": {
              "type": "array",
              "minItems": 15,
              "maxItems": 15,
              "prefixItems": [
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 1
                },
                {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 1
                }
              ],
              "items": false,
              "description": "点数 3..A,2,小王,大王 的张数"
            },
            "minItems": 2,
            "maxItems": 2
          },
          "firstSeat": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1
          },
          "currentSeat": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1
          },
          "isFirstMove": {
            "type": "boolean"
          },
          "targetMove": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Move"
              },
              {
                "type": "null"
              }
            ]
          },
          "lastPlaySeat": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "consecutivePasses": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "schemaVersion",
          "rulesetVersion",
          "hands",
          "firstSeat",
          "currentSeat",
          "isFirstMove",
          "targetMove",
          "lastPlaySeat",
          "consecutivePasses"
        ],
        "description": "hands[0]=USER，hands[1]=BOT；CLASSIC_V1 固定不支持四带二。"
      },
...
      "IdRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string",
            "maxLength": 128
          },
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "field",
          "reason"
        ]
      },
      "ErrorData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "minItems": 0,
            "maxItems": 100
          },
          "currentVersion": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 2147483647
              },
              {
                "type": "null"
              }
            ]
          },
          "retryAfterSeconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 2147483647
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": []
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "integer",
            "minimum": 40000,
            "maximum": 59999
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorData"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
      "Budget": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "timeLimitMs": {
            "type": "integer",
            "minimum": 100,
            "maximum": 60000,
            "default": 30000
          },
          "nodeLimit": {
            "type": "integer",
            "minimum": 1000,
            "maximum": 5000000,
            "default": 1000000
          },
          "memoryLimitMb": {
            "type": "integer",
            "minimum": 16,
            "maximum": 128,
            "default": 64
          }
        },
        "required": [
          "timeLimitMs",
          "nodeLimit",
          "memoryLimitMb"
        ]
      },
      "PuzzleSource": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PUZZLE_VERSION"
            ]
          },
          "puzzleVersionId": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          }
        },
        "required": [
          "type",
          "puzzleVersionId"
        ]
      },
      "GameSource": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "GAME"
            ]
          },
          "gameId": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "expectedStateVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          }
        },
        "required": [
          "type",
          "gameId",
          "expectedStateVersion"
        ]
      },
      "CustomSource": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "CUSTOM"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/GameState"
          }
        },
        "required": [
          "type",
          "state"
        ]
      },
      "SolveCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "source": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PuzzleSource"
              },
              {
                "$ref": "#/components/schemas/GameSource"
              },
              {
                "$ref": "#/components/schemas/CustomSource"
              }
            ],
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "PUZZLE_VERSION": "#/components/schemas/PuzzleSource",
                "GAME": "#/components/schemas/GameSource",
                "CUSTOM": "#/components/schemas/CustomSource"
              }
            }
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          }
        },
        "required": [
          "source"
        ]
      },
      "SolveStats": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "elapsedMs": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "queueTimeMs": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "visitedNodes": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "cacheHits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "elapsedMs",
          "queueTimeMs",
          "visitedNodes",
          "cacheHits"
        ]
      },
      "WinningLine": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "lineNo": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "moves": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Move"
            },
            "minItems": 0,
            "maxItems": 256
          },
          "verified": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "lineNo",
          "moves",
          "verified"
        ]
      },
      "SolveResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "proofStatus": {
            "type": "string",
            "enum": [
              "PROVEN",
              "UNKNOWN"
            ]
          },
          "winnerSide": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "USER",
                  "BOT"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "proofValue": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  -1,
                  1
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "strategyStatus": {
            "type": "string",
            "enum": [
              "NONE",
              "PARTIAL",
              "COMPLETE"
            ]
          },
          "rootNodeId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 19,
                "pattern": "^[1-9][0-9]{0,18}$",
                "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendedMove": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Move"
              },
              {
                "type": "null"
              }
            ]
          },
          "winningLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WinningLine"
            },
            "minItems": 0,
            "maxItems": 5
          },
          "solverVersion": {
            "type": "string",
            "maxLength": 64
          },
          "rulesetVersion": {
            "type": "string",
            "enum": [
              "CLASSIC_V1"
            ]
          },
          "stats": {
            "$ref": "#/components/schemas/SolveStats"
          }
        },
        "required": [
          "id",
          "proofStatus",
          "winnerSide",
          "proofValue",
          "strategyStatus",
          "rootNodeId",
          "recommendedMove",
          "winningLines",
          "solverVersion",
          "rulesetVersion",
          "stats"
        ],
        "description": "UNKNOWN 必须 winnerSide/proofValue=null；COMPLETE 必须 PROVEN；winningLines 最多展示5条已验证路线，不能替代完整多分支证明。"
      },
      "SolveTask": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "COMPLETED",
              "CANCELLED",
              "FAILED"
            ]
          },
          "purpose": {
            "type": "string",
            "enum": [
              "ANALYSIS",
              "HINT"
            ]
          },
          "effectiveBudget": {
            "$ref": "#/components/schemas/Budget"
          },
          "stateHash": {
            "type": "string",
            "maxLength": 64
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SolveResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "terminationReason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PROVED",
                  "TIME_LIMIT",
                  "NODE_LIMIT",
                  "MEMORY_LIMIT",
                  "CANCELLED",
                  "ERROR"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "stats": {
            "$ref": "#/components/schemas/SolveStats"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "UTC ISO-8601，输出以 Z 结束"
          },
          "startTime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "UTC ISO-8601，输出以 Z 结束"
              },
              {
                "type": "null"
              }
            ]
          },
          "finishTime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "UTC ISO-8601，输出以 Z 结束"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "taskStatus",
          "purpose",
          "effectiveBudget",
          "stateHash",
          "result",
          "terminationReason",
          "stats",
          "createTime",
          "startTime",
          "finishTime"
        ]
      },
      "TaskSearchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pageNum": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000,
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 20
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "COMPLETED",
              "CANCELLED",
              "FAILED"
            ]
          }
        },
        "required": []
      },
      "TaskPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SolveTask"
            },
            "minItems": 0,
            "maxItems": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "pageNum": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "items",
          "total",
          "pageNum",
          "pageSize"
        ]
      },
...
      "SolveTaskResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/SolveTask"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
      "TaskPageResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/TaskPage"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
    }
  }
}